split CompletionAction into a separate file
