+++ src/System.CommandLine.Tests/CompletionTests.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.CommandLine.Completions;
-using System.CommandLine.Parsing;
 using System.CommandLine.Tests.Utility;
 using System.IO;
 using System.Linq;
